FIX: 회원가입 기본 설정 모달 내용만 바뀌도록 수정
